requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
opencv-contrib-python=4.5.3.56=pypi_0
opencv-python=4.5.4.60=pypi_0
opencv-python-headless=4.5.3.56=pypi_0
pillow=8.4.0=py37h5aabda8_0
keras=2.4.3=py_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
mtcnn=0.1.1=pypi_0
pandas=1.3.4=py37h8c16a72_0
pyttsx3=2.90=pypi_0
unicodecsv=0.14.1=py37_0
